requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
streamlit==1.29.0
flask_cors